+++ CHANGELOG.md
@@ -15,6 +15,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   the suggestion to enable the `hack_reaper_update_display` workaround for
   REAPER when it is not already enabled.
 
+### Fixed
+
+- Changed the way keyboard input focus works to also allow keyboard input in
+  _REAPER_. Please let me know if this causes any issues elsewhere!
+
 ## [1.6.1] - 2020-09-28
 
 ### Fixed